rag_project/urls.py
# rag_project/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from django.views.generic import TemplateView
from authentication import CustomTokenObtainPairView
from rest_framework_simplejwt.views import TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/chat/', include('chat.urls')),
    path('api/knowledge-base/', include('knowledge_base.urls')),
    path('api/models/', include('model_manager.urls')),
    
    # 用户管理与JWT认证
    path('api/token/', CustomTokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/users/', include('users.urls')),
    
    # 添加React前端的入口点
    path('', TemplateView.as_view(template_name='index.html')),
    
    # 添加ws路径用于调试
    path('ws/', TemplateView.as_view(template_name='index.html')),
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)