requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2txt==0.9
emoji==2.14.1
et_xmlfile==1.0.0
eval_type_backport==0.2.2
faiss-cpu==1.10.0
filetype==1.2.0
frozenlist==1.6.0
greenlet==3.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.32
lxml==5.3.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.5
olefile==0.47
openai==1.75.0
openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.9.0
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.33.0
urllib3==2.4.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0
PyMuPDF==1.22.0
pdfplumber==0.10.0 
pdf2image==1.16.0  
pytesseract==0.3.10  
uuid==1.30  
textract==1.6.5  
tiktoken==0.5.0  
threading  
@ant-design/icons@6.0.0
antd@5.24.8
axios@1.8.4
katex@0.16.22
prismjs@1.30.0
react@19.1.0
react-dom@19.1.0
react-markdown@10.1.0
react-router-dom@7.5.1
react-scripts@5.0.1
react-syntax-highlighter@15.6.1
uuid@9.0.0
zustand@5.0.3